Increasing the thread stack space
While working with the Sematext folks to get SPM running on our QA
environment, we noticed that SPM would not start with Java 6. After
some investigation, the Sematext folks discovered that SPM would not
start because of the low -Xss setting. This change is to address this.
(see https://github.com/karmi/cookbook-elasticsearch/issues/25)